add auth provider
